Register become mentor
update

+++ src/java/controller/ActionAdminRegisterBecomeMentor.java
@@ -2,7 +2,6 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package controller;
 
 import java.io.IOException;
@@ -20,8 +19,8 @@ import model.*;
  * @author Admin
  */
 public class ActionAdminRegisterBecomeMentor extends HttpServlet {
-   
-  protected void doGet(HttpServletRequest request, HttpServletResponse response)
+    
+    protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         PrintWriter out = response.getWriter();
@@ -31,45 +30,51 @@ public class ActionAdminRegisterBecomeMentor extends HttpServlet {
         AccountDAO aDAO = new AccountDAO();
         MenteeDAO meDAO = new MenteeDAO();
         MentorSkillDAO msDAO = new MentorSkillDAO();
+        RequestDAO rDAO = new RequestDAO();
         String action = request.getParameter("action");
-        Account acc = (Account)request.getSession().getAttribute("acc");
+        Account acc = (Account) request.getSession().getAttribute("acc");
         int idMentee = Integer.parseInt(request.getParameter("ID"));
-        if(action.equals("reject")){
-          lDAO.delete(idMentee);
-          String content = "If you failed to apply becoming a mentor, you can try again when your CV is better";
-          nDAO.insertNoficationDAO(acc.getID(),idMentee , content, 1);
-          request.getRequestDispatcher("ListBecomeMentor.jsp").forward(request, response);
-        }else{
-            SkillDAO sDAO = new SkillDAO();
-            List<Skill> listSkill = sDAO.getListSkill();
-            CVToBecomeMentor cv = lDAO.getCVBecomeMentorByID(idMentee);
-            String[] skillMentor = cv.getSkill().split(", ");
-            List<Integer>listIDSkill = new ArrayList<>();
-            for(String name:skillMentor){
-               for(Skill s:listSkill){
-                   if(s.getName().equals(name)){
-                       listIDSkill.add(s.getID());
-                   }
-               } 
+        if (action.equals("reject")) {
+            lDAO.delete(idMentee);
+            String content = "If you failed to apply becoming a mentor, you can try again when your CV is better";
+            nDAO.insertNoficationDAO(acc.getID(), idMentee, content, 1);
+            request.getRequestDispatcher("ListBecomeMentor.jsp").forward(request, response);
+        } else {            
+            if (rDAO.IsExistRequestAccepted(idMentee)) {
+                request.setAttribute("Mentee is having a meeting, can't become mentor now", out);
+            } else {
+                SkillDAO sDAO = new SkillDAO();
+                List<Skill> listSkill = sDAO.getListSkill();
+                CVToBecomeMentor cv = lDAO.getCVBecomeMentorByID(idMentee);
+                String[] skillMentor = cv.getSkill().split(", ");
+                List<Integer> listIDSkill = new ArrayList<>();
+                for (String name : skillMentor) {
+                    for (Skill s : listSkill) {
+                        if (s.getName().equals(name)) {
+                            listIDSkill.add(s.getID());
+                        }
+                    }                    
+                }
+                rDAO.deleteRequestByIDMentee(idMentee);
+                aDAO.updateRole(idMentee);
+                meDAO.delete(idMentee);
+                mDAO.insertMentor(idMentee, cv.getIntro(), cv.getEx());
+                for (int i : listIDSkill) {
+                    msDAO.insertMentorSkill(idMentee, i);
+                }
+                lDAO.delete(idMentee);
+                String content = "Congratulations on becoming a mentor, I hope you will try your best to help as much as possible mentees";
+                nDAO.insertNoficationDAO(acc.getID(), idMentee, content, 1);                
             }
-         aDAO.updateRole(idMentee);
-         meDAO.delete(idMentee);
-         mDAO.insertMentor(idMentee, cv.getIntro(), cv.getEx());
-         for(int i: listIDSkill){
-           msDAO.insertMentorSkill(idMentee, i);
-         }
-         lDAO.delete(idMentee);
-         String content = "Congratulations on becoming a mentor, I hope you will try your best to help as much as possible mentees";
-         nDAO.insertNoficationDAO(acc.getID(), idMentee, content, 1);
-         request.getRequestDispatcher("ListBecomeMentor.jsp").forward(request, response);
+            request.getRequestDispatcher("ListBecomeMentor.jsp").forward(request, response);
         }
     }
-
+    
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         PrintWriter out = response.getWriter();
-     
+        
     }
-
+    
 }
